+++ src/Things/Structures/Residence.h
@@ -4,6 +4,14 @@
 
 #include "../../Constants.h"
 
+
+/**
+ * \class	Residence
+ * \brief	Base Residential structure.
+ * 
+ * Implements the base Residence structures. Upgraded residences should derive
+ * from this class and override the definition of mCapacity.
+ */
 class Residence : public Structure
 {
 public:
@@ -16,15 +24,15 @@ public:
 		requiresCHAP(true);
 	}
 
-
 	virtual ~Residence()
 	{}
 
-protected:
 
+	int capacity() const { return mCapacity; }
+
+protected:
 	virtual void think()
-	{
-	}
+	{}
 
 	virtual void defineResourceInput()
 	{
@@ -34,6 +42,7 @@ protected:
 	virtual void defineResourceOutput()
 	{}
 
-private:
+protected:
+	int		mCapacity = 25;		/**< Override this value in derived residences.*/
 
 };