+++ src/UI/PopulationPanel.h
@@ -5,9 +5,9 @@
 class PopulationPanel: public Control
 {
 public:
-	PopulationPanel(): mIcons("ui/icons.png"), mPopulation(nullptr), mMorale(nullptr), mPreviousMorale(nullptr)
+	PopulationPanel(): mIcons("ui/icons.png")
 	{
-		size(125, 220);
+		size(160, 220);
 
 		mSkin.push_back(Image("ui/skin/window_top_left.png"));
 		mSkin.push_back(Image("ui/skin/window_top_middle.png"));
@@ -26,37 +26,50 @@ public:
 	void morale(int* m) { mMorale = m; }
 	void old_morale(int* m) { mPreviousMorale = m; }
 
+	void residential_capacity(int m) { mResidentialCapacity = m; }
+
 	virtual void update()
 	{
 		Renderer& r = Utility<Renderer>::get();
-
-		//r.drawBoxFilled(rect(), 0, 0, 0);
 		r.drawImageRect(rect().x(), rect().y(), rect().width(), rect().height(), mSkin);
 
 		r.drawText(font(), string_format("Morale: %i", *mMorale), positionX() + 5, positionY() + 5, 255, 255, 255);
 		r.drawText(font(), string_format("Previous: %i", *mPreviousMorale), positionX() + 5, positionY() + 15, 255, 255, 255);
 
-		r.drawSubImage(mIcons, positionX() + 5, positionY() + 40, 0, 96, 32, 32);		// Infant
-		r.drawSubImage(mIcons, positionX() + 5, positionY() + 74, 32, 96, 32, 32);		// Student
-		r.drawSubImage(mIcons, positionX() + 5, positionY() + 108, 64, 96, 32, 32);		// Worker
-		r.drawSubImage(mIcons, positionX() + 5, positionY() + 142, 96, 96, 32, 32);		// Scientist
-		r.drawSubImage(mIcons, positionX() + 5, positionY() + 176, 128, 96, 32, 32);	// Retired
+		if (mResidentialCapacity == 0)
+		{
+			r.drawText(font(), string_format("Housing: %i / %i  (0%%)", mPopulation->size(), mResidentialCapacity), positionX() + 5, positionY() + 30, 255, 255, 255);
+		}
+		else
+		{
+			float capacity = (float)mPopulation->size() / (float)mResidentialCapacity;
+			r.drawText(font(), string_format("Housing: %i / %i  (%i%%)", mPopulation->size(), mResidentialCapacity, static_cast<int>(capacity * 100)), positionX() + 5, positionY() + 30, 255, 255, 255);
 
-		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_CHILD)), positionX() + 42, positionY() + 60, 255, 255, 255);
-		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_STUDENT)), positionX() + 42, positionY() + 92, 255, 255, 255);
-		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_WORKER)), positionX() + 42, positionY() + 124, 255, 255, 255);
-		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_SCIENTIST)), positionX() + 42, positionY() + 155, 255, 255, 255);
-		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_RETIRED)), positionX() + 42, positionY() + 188, 255, 255, 255);
+		}
+
+		r.drawSubImage(mIcons, positionX() + 5, positionY() + 45, 0, 96, 32, 32);		// Infant
+		r.drawSubImage(mIcons, positionX() + 5, positionY() + 79, 32, 96, 32, 32);		// Student
+		r.drawSubImage(mIcons, positionX() + 5, positionY() + 113, 64, 96, 32, 32);		// Worker
+		r.drawSubImage(mIcons, positionX() + 5, positionY() + 147, 96, 96, 32, 32);		// Scientist
+		r.drawSubImage(mIcons, positionX() + 5, positionY() + 181, 128, 96, 32, 32);	// Retired
+
+		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_CHILD)), positionX() + 42, positionY() + 65, 255, 255, 255);
+		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_STUDENT)), positionX() + 42, positionY() + 97, 255, 255, 255);
+		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_WORKER)), positionX() + 42, positionY() + 129, 255, 255, 255);
+		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_SCIENTIST)), positionX() + 42, positionY() + 160, 255, 255, 255);
+		r.drawText(font(), string_format("%i", mPopulation->size(Population::ROLE_RETIRED)), positionX() + 42, positionY() + 193, 255, 255, 255);
 	}
 
 protected:
 
 private:
 	Image		mIcons;
-
 	ImageList	mSkin;
 
-	Population*	mPopulation;
-	int*		mMorale;
-	int*		mPreviousMorale;
+	Population*	mPopulation = nullptr;
+
+	int			mResidentialCapacity = 0;
+
+	int*		mMorale = nullptr;
+	int*		mPreviousMorale = nullptr;
 };